Added a lot of things ehe

+++ app/Http/Controllers/MahasiswaController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Mahasiswa;
+use App\Models\MataKuliah;
 use Illuminate\Http\Request;
 
 
@@ -22,7 +23,7 @@ class MahasiswaController extends Controller
         return response()->json([
             'status' => 'success',
             'mahasiswa' => $mahasiswa
-        ]);
+        ], 200);
     }
 
     public function getMahasiswaByNim($nim)
@@ -38,7 +39,7 @@ class MahasiswaController extends Controller
                 "angkatan" => $mahasiswa->angkatan,
                 "matakuliah" => $mahasiswa->matakuliah
             ]
-        ]);
+        ], 200);
     }
 
     public function getMahasiswaByToken(Request $request)
@@ -46,29 +47,71 @@ class MahasiswaController extends Controller
         return response()->json([
             'status' => 'success',
             "mahasiswa" => $request->user
-        ]);
+        ], 200);
     }
 
-    public function AddMataKuliahToMahasiswa($nim, $mkId)
+    public function AddMataKuliahToMahasiswa(Request $request, $nim, $mkId)
     {
         $nimMhsw = $nim;
         $matakuliah = $mkId;
 
-        $mahasiswa = Mahasiswa::find($nimMhsw);
+        if($matakuliah > 5) {
+            return response()->json([
+                "status" => "error",
+                "message" => "Mata kuliah tidak ada"
+            ], 400);
+        }
+
+        $dataMahasiswa = Mahasiswa::with('MataKuliah')->where("nim", $request->nim)->first();
+        $matkulMahasiswa = $dataMahasiswa->matakuliah;
+        $isMatkulExist = false;
+        
+
+        for ($i=0; $i < count($matkulMahasiswa) ; $i++) { 
+            if ($matkulMahasiswa[$i]->id == $matakuliah ) {
+                $isMatkulExist = true;
+            }
+        }
+
+
+        if($isMatkulExist) {
+            return response()->json([
+                "status" => "error",
+                "message" => "Mata kuliah sudah terdaftar, silahkan pilih yang lain"
+            ], 200);
+        }
+
+        if($request->nim != $nimMhsw) {
+            return response()->json([
+                "status" => "error",
+                "message" => "Kredensial login berbeda"
+            ], 200);
+        }
+
+        $mahasiswa = Mahasiswa::find($request->nim);
         $mahasiswa->MataKuliah()->attach($matakuliah);
 
         return response()->json([
-            "mahasiswa" => $mahasiswa
-        ]);
+            "status" => "success",
+            "message" => "Mata kuliah berhasil ditambah"
+        ], 200);
     }
 
-    public function DeleteMataKuliahOnMahasiswa($nim, $mkId)
-    {
+    public function DeleteMataKuliahOnMahasiswa(Request $request, $nim, $mkId)
+    {   
+        if($request->nim != $nim) {
+            return response()->json([
+                "status" => "error",
+                "message" => "Kredensial login berbeda"
+            ], 400);
+        }
+
         $mahasiswa = Mahasiswa::find($nim);
         $mahasiswa->MataKuliah()->detach($mkId);
 
         return response()->json([
-            "success" => "Berhasil dihapus"
-        ]);
+            "status" => "success",
+            "message" => "Mata kuliah berhasil dihapus"
+        ], 200);
     }
 }
